fixing issue 1820
